requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
cffi==2.0.0
pycparser==3.0
pycryptodome==3.23.0
pysimplegui==4.60.5.1
